Add 1 test for adaptor to switch from v1 to v2
